# Conflicts:
#	build.gradle
#	core/src/main/assets/interfaces/icons.png
#	core/src/main/assets/messages/items/items.properties
#	core/src/main/assets/messages/scenes/scenes.properties
#	core/src/main/assets/messages/windows/windows.properties
#	core/src/main/assets/sprites/items.png
#	core/src/main/assets/sprites/shaman.png
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/Assets.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/Badges.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/Actor.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/Char.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Bleeding.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Burning.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Corrosion.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Corruption.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Frost.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Hunger.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Ooze.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Poison.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/Hero.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/ArmoredStatue.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/DM100.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Eye.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/GnollSapper.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Statue.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/VaultMob.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Warlock.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/YogDzewa.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/npcs/Imp.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/effects/MagicMissile.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/Generator.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Viscosity.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/DriedRose.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/HornOfPlenty.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/MasterThievesArmband.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/TalismanOfForesight.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/bombs/ArcaneBomb.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/quest/EscapeCrystal.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/stones/StoneOfAugmentation.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/ChaoticCenser.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/Wand.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfBlastWave.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/Weapon.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/enchantments/Corrupting.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/MissileWeapon.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/AdrenalineDart.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/CleansingDart.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/HealingDart.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/HolyDart.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/journal/Catalog.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/CityLevel.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/VaultLevel.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/BlacksmithRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/VaultLasersRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/VaultLongRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/VaultSimpleEnemyTreasureRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultBookcaseTreasureRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultFlamePathRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultLaserTreasureRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultManyScansRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultMultipleEnemyTreasureRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/quest/vault/treasure/VaultSingleEnemyTreasureRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/secret/RatKingRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/standard/entrance/HallwayEntranceRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/standard/entrance/StatuesEntranceRoom.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/AboutScene.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/ChangesScene.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/sprites/ItemSpriteSheet.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/sprites/ShamanSprite.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ui/Icons.java

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/ArmoredStatue.java
@@ -30,6 +30,7 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.armor.Armor;
 import com.shatteredpixel.shatteredpixeldungeon.items.armor.ClothArmor;
+import com.shatteredpixel.shatteredpixeldungeon.items.weapon.enchantments.Corrupting;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.CharSprite;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.StatueSprite;
@@ -119,8 +120,11 @@ public class ArmoredStatue extends Statue {
 
 	@Override
 	public void die( Object cause ) {
-		armor.identify(false);
-		Dungeon.level.drop( Challenges.process(armor), pos ).sprite.drop();
+		//silently duplicated in this case, so don't drop original armor
+		if (buff(Corrupting.CorruptingTracker.class) == null) {
+			armor.identify(false);
+			Dungeon.level.drop(Challenges.process(armor), pos).sprite.drop();
+		}
 		super.die( cause );
 	}
